Refactor frontend
Add loading indicator when checking session
Separate home page and assignment view from main App component
