+++ Frontend/package.json
@@ -12,7 +12,8 @@
   "dependencies": {
     "axios": "^1.6.5",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-loader-spinner": "^6.1.6"
   },
   "devDependencies": {
     "@types/react": "^18.2.43",
