+++ Frontend/src/css/App.css
@@ -10,7 +10,6 @@ h3 {
 body {
   margin: 0;
   padding: 0;
-  overflow: hidden;
 }
 
 .floor {
